+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added an optional `internalOptions` parameter to `getClient` with an `addDefaultAcceptHeader` flag (defaults to `true`). When set to `false`, the client no longer adds a default `Accept: application/json` request header to operations that do not otherwise specify one, allowing operations that expect no response body to omit the `Accept` header. This option is intended for use by generated clients. [PR #39291](https://github.com/Azure/azure-sdk-for-js/pull/39291)
+- Added an optional `internal` property to `ClientOptions` for `getClient`, carrying a `noDefaultAcceptHeader` flag (defaults to `false`). When set to `true`, the client no longer adds a default `Accept: application/json` request header to operations that do not otherwise specify one, allowing operations that expect no response body to omit the `Accept` header. This option is intended for use by generated clients. [PR #39291](https://github.com/Azure/azure-sdk-for-js/pull/39291)
 
 ### Breaking Changes